requirements.txt
easydict==1.10
Flask==2.3.2
numpy==1.24.1
Pillow==9.3.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tqdm==4.65.0
wandb==0.15.4
